# Implement option in WCS class to preserve the original units