--- a/astropy/wcs/include/astropy_wcs/wcslib_wrap.h
+++ b/astropy/wcs/include/astropy_wcs/wcslib_wrap.h
@@ -10,8 +10,24 @@
 extern PyTypeObject PyWcsprmType;
 
 typedef struct {
+
   PyObject_HEAD
   struct wcsprm x;
+
+  // WCSLIB converts units to SI (for example nanometers for a specrtal axis to
+  // meters, or arcseconds for a celestial frame to degrees). There is no easy
+  // way to turn off this behavior in WCSLIB itself, but since we expose WCSLIB
+  // through a wrapper layer, we can optionally do conversions on-the-fly to
+  // make it look to the user as if the units are the original ones. This is
+  // controlled via the preserve_units option which can be 0 (default WCSLIB
+  // behavior) or 1 (preserve original units). The original_cunit array is used
+  // to store the original CUNIT values, while the unit_scaling array contains
+  // the multiplicative scaling required to convert the units from the original
+  // units (e.g. arcsec) to the internal WCSLIB units (e.g. deg)
+  int    preserve_units;
+  char   (*original_cunit)[72];
+  double *unit_scaling;
+
 } PyWcsprm;
 
 int _setup_wcsprm_type(PyObject* m);