--- a/astropy/wcs/wcs.py
+++ b/astropy/wcs/wcs.py
@@ -347,6 +347,12 @@ class WCS(FITSWCSAPIMixin, WCSBase):
         `WCS.fix` for more information about this parameter.  Only
         effective when ``fix`` is `True`.
 
+    preserve_units : str, optional
+        By default, some units are converted to SI, for example spectral
+        axes in units of nm might be converted to m, and celestial axes
+        in units of arcsec might be converted to deg. If ``preserve_units``
+        is set to `True`, the original units will be preserved.
+
     Raises
     ------
     MemoryError
@@ -418,6 +424,7 @@ class WCS(FITSWCSAPIMixin, WCSBase):
         fix=True,
         translate_units="",
         _do_set=True,
+        preserve_units=False,
     ):
         close_fds = []
 
@@ -430,7 +437,13 @@ class WCS(FITSWCSAPIMixin, WCSBase):
         if header is None:
             if naxis is None:
                 naxis = 2
-            wcsprm = _wcs.Wcsprm(header=None, key=key, relax=relax, naxis=naxis)
+            wcsprm = _wcs.Wcsprm(
+                header=None,
+                key=key,
+                relax=relax,
+                naxis=naxis,
+                preserve_units=preserve_units,
+            )
             self.naxis = wcsprm.naxis
             # Set some reasonable defaults.
             det2im = (None, None)
@@ -502,6 +515,7 @@ class WCS(FITSWCSAPIMixin, WCSBase):
                     colsel=colsel,
                     warnings=False,
                     hdulist=fobj,
+                    preserve_units=preserve_units,
                 )
                 if naxis is not None:
                     try:
@@ -541,6 +555,7 @@ class WCS(FITSWCSAPIMixin, WCSBase):
                     keysel=keysel_flags,
                     colsel=colsel,
                     hdulist=fobj,
+                    preserve_units=preserve_units,
                 )
             except _wcs.NoWcsKeywordsFoundError:
                 # The header may have SIP or distortions, but no core
@@ -554,6 +569,7 @@ class WCS(FITSWCSAPIMixin, WCSBase):
                         keysel=keysel_flags,
                         colsel=colsel,
                         hdulist=fobj,
+                        preserve_units=preserve_units,
                     )
                 else:
                     raise