### Description

A common issue users run into when using the ``astropy.wcs`` class is that WCSLIB will convert units to SI. For instance, celestial WCS in arcseconds will be converted to degrees, and spectral axes in GHz and nm will be converted to m. There isn't much we can do to stop WCSLIB from doing this - however I have realized that since we never expose the WCSLIB objects directly, but instead have our own C wrapper class, we can make that wrapper class remember any initial units and do on-the-fly conversions for all appropriate methods/properties.

As an example:

```python
In [1]: from astropy.io import fits
   ...: 

In [2]: header = fits.Header()
   ...: header['CTYPE1'] = 'RA---TAN'
   ...: header['CTYPE2'] = 'DEC--TAN'
   ...: header['CUNIT1'] = 'arcsec'
   ...: header['CUNIT2'] = 'arcsec'
   ...: header['CRVAL1'] = 20
   ...: header['CRVAL2'] = 20
   ...: header['CRPIX1'] = 1
   ...: header['CRPIX2'] = 1
   ...: header['CDELT1'] = -1
   ...: header['CDELT2'] = 1
   ...: 

In [3]: from astropy.wcs import WCS
   ...: 

In [4]: wcs = WCS(header)

In [5]: wcs.wcs.cunit
Out[5]: ['deg', 'deg']

In [6]: wcs.wcs.crval
Out[6]: array([0.00555556, 0.00555556])

In [8]: wcs.all_pix2world(1, 1, 0)
Out[8]: [array(0.00527778), array(0.00583333)]

In [9]: wcs2 = WCS(header, preserve_units=True)

In [10]: wcs2.wcs.cunit
Out[10]: ['arcsec', 'arcsec']

In [11]: wcs2.wcs.crval
Out[11]: array([20., 20.])

In [13]: wcs2.all_pix2world(1, 1, 0)
Out[13]: [array(18.99999999), array(21.)]

```

This is not yet complete but I am putting it up already so that people who are interested can try it out.

Remaining TODOs:

```
* [ ] Extend support to headers with CD matrices
* [ ] Fix/check support for printing the WCS
* [ ] Check that we can dynamically change the units on the WCS and that things behave correctly
* [ ] Expand test suite
* [ ] Write documentation
* [ ] Add news fragment

This will need to be rebased once https://github.com/astropy/astropy/pull/18231 is merged

- [ ] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.