@astrofrog 2025-06-25
I'm seeing a test failure in ``test_wtbarr_print`` which is a test unrelated to the changes here but what is happening is that the ``print_contents()`` call there includes some output from the new test I added in this PR, so somehow something isn't getting cleared in memory.