@mcara 2025-07-14
@astrofrog While I review this, may I ask whether you had contacted Mark to see if this could be fixed in WCSLIB? It seems it would be a much more natural place for the fix as it would avoid unnecessary copying of `Wcsprm` structure. It seems unnatural/unexpected for WCSLIB to (kind of) arbitrarily convert units in the `Wcsprm` structure.

Section 2.3 of Paper I says:

> 2.3. Keyword value units
The original FITS paper (Wells et al. 1981) assumed that the units along each axis could be implied simply by the contents of the CTYPEi keyword. This has not turned out to be true in general. Therefore, we propose adding a new indexed, keyword CUNITi (character-valued) with which the units of CRVALi and CDELTi may be specified. Restrictions on the nature and range of units, if any, will be determined by the agreements applying to the specific axes. If they are not so limited, units should conform with the recommendations of the IAU Style Manual (McNally 1988).

Although in Paper II (section 1) it says:
> Wells et al. (1981) established that all angles in FITS were to be measured in degrees and this has been entrenched by the AIPS convention and confirmed in the IAU-endorsed FITS standard (Hanisch et al. 2001). Paper I introduced the CUNITia keyword to define the units of CRVALia and CDELTia. **Accordingly, we require CUNITia = 'deg' for the celestial CRVALia and CDELTia, whether given explicitly or not. Consequently, the (x; y) coordinates in the plane of projection are measured in degrees.** For consistency, we use degree measure for native and celestial spherical coordinates and for all other angular measures in this paper.

This is confusing (to me). So I looked at the Hanisch et al. 2001 paper, section 5.3, and it says:

> The units of all FITS header keyword values, with the exception of measurements of angles, should conform with the recommendations in the IAU Style Manual (McNally 1988). For angular measurements given as floating point values and specified with reserved keywords, degrees are the recommended units** (with the units, if specified, given as ’deg’).

recommended != required. I think it is worth asking why WCSLIB chose to behave like that and whether it is a bug and regardless of bug or not - can't this be handled there?

While I think it is worth contacting Mark Calabretta about this, I think we can go ahead and implement this here.