@mcara 2025-07-15
I also wonder how this works with the `-TAB` non-linear algorithm (see section 6 in Paper III - spectral). Maybe add an example using -TAB?