@astrofrog 2025-07-15
> What is the performance cost for this for the case of a single coordinate transformation (a pair of coordinates) and for a large array of coordinates?

Using the 5-D WCS from the test added in this PR (``wcs_default`` is without ``preserve_units=True``, and ``wcs_preserve`` is with), and doing a lot of repeated measurements, I get:

**Before this PR**

wcs_default.all_pix2world(2, 3, 8, 1, 2, 0)        0.014 +/- 0.002 ms
wcs_default.all_pix2world(coords, 0)               1.012 +/- 0.076 ms
wcs_default.all_world2pix(2, 3, 8, 1, 2, 0)        0.028 +/- 0.008 ms
wcs_default.all_world2pix(coords, 0)               1.178 +/- 0.146 ms

**With this PR**

```
wcs_default.all_pix2world(2, 3, 8, 1, 2, 0)        0.013 +/- 0.001 ms
wcs_default.all_pix2world(coords, 0)               1.112 +/- 0.017 ms
wcs_default.all_world2pix(2, 3, 8, 1, 2, 0)        0.017 +/- 0.001 ms
wcs_default.all_world2pix(coords, 0)               1.024 +/- 0.057 ms
wcs_preserve.all_pix2world(2, 3, 8, 1, 2, 0)       0.022 +/- 0.001 ms
wcs_preserve.all_pix2world(coords, 0)              1.204 +/- 0.087 ms
wcs_preserve.all_world2pix(2, 3, 8, 1, 2, 0)       0.028 +/- 0.001 ms
wcs_preserve.all_world2pix(coords, 0)              0.221 +/- 0.013 ms
```

Time times for ``wcs_default`` are consistent within errors, so no performance impact for the default case.

With preserve units, it seems pix2world is a little slower for ``pix2world``, which is to be expected, but I am a bit surprised by ``world2pix`` actually being faster, so I will try and investigate that.