@astrofrog 2025-07-15
@mcara - I am having issues getting a simple -TAB example to work. I would have thought the following would work:

```python
import numpy as np
from astropy.io import fits
from astropy.wcs import WCS

flux = np.random.rand(100)

wav = np.linspace(4000, 5000, 100)

primary_hdu = fits.PrimaryHDU(flux)

col = fits.Column(name='WAVE', format='D', array=wav)
tab_hdu = fits.BinTableHDU.from_columns([col])
tab_hdu.name = 'WCS-TAB'

header = primary_hdu.header
header['WCSAXES'] = 1
header['CTYPE1'] = 'WAVE-TAB'
header['CRPIX1'] = 1
header['CDELT1'] = 1
header['CRVAL1'] = 0
header['CUNIT1'] = 'mm'
header['PS1_0'] = 'WCS-TAB'
header['PS1_1'] = 'WAVE'
header['PV1_1'] = 1

hdul = fits.HDUList([primary_hdu, tab_hdu])

wcs_default = WCS(hdul[0].header, hdul)

print(wcs_default.wcs_pix2world(0, 0))
print(wcs_default.wcs_pix2world(1, 0))
print(wcs_default.wcs_pix2world(2, 0))
print(wcs_default.wcs_pix2world(3, 0))
```

This outputs:

```
[array(nan)]
[array(4000.)]
[array(nan)]
[array(nan)]
```

Only pixel values between 1 and 2 seem to return a valid world coordinate. Have I made a mistake in setting up my FITS file?