@mcara 2025-07-15
```python
In [2]: wp = wcs_preserve.all_world2pix(world_coords_native, 0)

In [3]: wnp = wcs_default.all_world2pix(world_coords_si, 0)

In [4]: wp
Out[4]:
array([[-1.        ,         inf, -1.        , -4.        ,  2.71759582],
       [-1.        ,         inf, -1.        , -4.        ,  2.79517881],
       [-1.        ,         inf, -1.        , -4.        ,  2.82613895],
       ...,
       [-1.        ,         inf, -1.        , -4.        ,  2.73230782],
       [-1.        ,         inf, -1.        , -4.        ,  2.81918181],
       [-1.        ,         inf, -1.        , -4.        ,  2.81773117]],
      shape=(10000, 5))

In [5]: wnp
Out[5]:
array([[-0.78075305, -0.51671735, -0.55508113, -3.88556918,  2.71759582],
       [-0.83539454, -0.5772501 , -0.50062671, -3.97816059,  2.79517881],
       [-0.94703594, -0.35614207, -0.8485017 , -3.50202099,  2.82613895],
       ...,
       [-0.81628231, -0.54181075, -0.96587219, -3.10801576,  2.73230782],
       [-0.83270913, -0.58271578, -0.88742555, -3.84659959,  2.81918181],
       [-0.99437225, -0.38951446, -0.72503993, -3.45589686,  2.81773117]],
      shape=(10000, 5))

In [6]: wcs_default
Out[6]:
WCS Keywords

Number of WCS axes: 5
CTYPE : 'RA---TAN' 'FREQ' 'DEC--TAN' 'WAVE' 'OFFSET'
CRVAL : 0.0011111111111111111 5000000000.0 0.0016666666666666666 7.000000000000001e-09 8.0
CRPIX : 1.0 2.0 3.0 4.0 5.0
PC1_1 PC1_2 PC1_3 PC1_4 PC1_5  : 1.0 0.0 0.0 0.0 0.0
PC2_1 PC2_2 PC2_3 PC2_4 PC2_5  : 0.0 1.0 0.0 0.0 0.0
PC3_1 PC3_2 PC3_3 PC3_4 PC3_5  : 0.0 0.0 1.0 0.0 0.0
PC4_1 PC4_2 PC4_3 PC4_4 PC4_5  : 0.0 0.0 0.0 1.0 0.0
PC5_1 PC5_2 PC5_3 PC5_4 PC5_5  : 0.0 0.0 0.0 0.0 1.0
CDELT : 0.0011111111111111111 3000000000.0 0.0005555555555555556 1e-09 6.0
NAXIS : 0  0
```

You can see that something is not working well here. Only the `'OFFSET'` column looks good. Also, since there are no distortions, `all_world2pix` should revert to `wcs_world2pix` (after a minor overhead).