@astrofrog 2025-07-15
Oh I see the issue, it looks like I am modifying the input world array in-place - oops!