@mcara 2025-07-15
@astrofrog So maybe add a roundtrip test (not a separate test - an `assert` would be enough) to detect these errors